+++ GREATClient/BaseClass/DrawableImage.cs
@@ -88,7 +88,7 @@ namespace GREATClient.BaseClass
 
 			batch.Begin();
 
-			batch.Draw(Texture,GetAbsolutePosition(),SourceRectangle,Tint * Alpha,(float)Orientation,
+			batch.Draw(Texture,GetAbsolutePosition(),SourceRectangle,Tint * Alpha, Orientation,
 			           RelativeOrigin * new Vector2(Width, Height),Scale,Effects,0);
 
 			batch.End();
